+++ CHANGELOG.md
@@ -12,3 +12,4 @@
   - 把 InputItem clear icon 设置为 x3s
   - 增加 `itemStyle?: ViewStyle` 属性
   - textAlign 增加 'right' 选项
+  - InputItem clear 增加 textAlign 值不等于 left 才显示的条件